Consolidate file into duke package
