+++ src/main/java/duke/Deadlines.java
@@ -1,8 +1,10 @@
+package duke;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
 public class Deadlines extends Task {
-    LocalDate deadLine;
+    protected LocalDate deadLine;
 
     public Deadlines(boolean isDone, String eventName, String deadLine) {
         super(isDone, eventName, "D");
